refactor(import): improve readability and move helper functions to import_utils
